+++ src/testRunners/jasmine/Jasmine2Reporter.js
@@ -79,7 +79,7 @@ class Jasmine2Reporter {
 
     specResult.failedExpectations.forEach(failed => {
       let message;
-      if (!failed.matcherName) {
+      if (!failed.matcherName && failed.stack) {
         message = this._formatter.formatException(failed.stack);
       } else {
         message = this._formatter.formatMatchFailure(failed);