+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 * Fixed `moduleNameMapper` config option when used with paths.
 * Added `jest-util`, general code cleanup.
 * Fixed an error with Jasmine 2 and tests that `throw 'string errors'`.
+* Fixed issues with unmocking symlinked module names.
 
 ## jest-cli 0.9.2, babel-jest 9.0.3